backend/pyproject.toml
[project]
name = "todo-backend"
version = "1.0.0"
description = "Todo Full-Stack Web Application - Backend API"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.109.0",
    "sqlmodel>=0.0.14",
    "bcrypt>=4.1.2",
    "pyjwt>=2.8.0",
    "pydantic[email]>=2.1.0",
    "pydantic-settings>=2.1.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "psycopg2-binary>=2.9.9",
]

[dependency-groups]
dev = []
